+++ cmd/kube-proxy/app/server.go
@@ -62,6 +62,7 @@ import (
 	"k8s.io/component-base/metrics/legacyregistry"
 	"k8s.io/component-base/version"
 	"k8s.io/component-base/version/verflag"
+	utilipset "k8s.io/component-helpers/network/utils/ipset"
 	utiliptables "k8s.io/component-helpers/network/utils/iptables"
 	"k8s.io/klog/v2"
 	"k8s.io/kube-proxy/config/v1alpha1"
@@ -83,7 +84,6 @@ import (
 	proxyutil "k8s.io/kubernetes/pkg/proxy/util"
 	"k8s.io/kubernetes/pkg/util/filesystem"
 	utilflag "k8s.io/kubernetes/pkg/util/flag"
-	utilipset "k8s.io/kubernetes/pkg/util/ipset"
 	utilipvs "k8s.io/kubernetes/pkg/util/ipvs"
 	"k8s.io/kubernetes/pkg/util/oom"
 	"k8s.io/utils/exec"
