+++ pkg/proxy/ipvs/proxier_test.go
@@ -35,6 +35,8 @@ import (
 	"k8s.io/apimachinery/pkg/util/sets"
 	utilfeature "k8s.io/apiserver/pkg/util/feature"
 	featuregatetesting "k8s.io/component-base/featuregate/testing"
+	utilipset "k8s.io/component-helpers/network/utils/ipset"
+	ipsettest "k8s.io/component-helpers/network/utils/ipset/testing"
 	utiliptables "k8s.io/component-helpers/network/utils/iptables"
 	iptablestest "k8s.io/component-helpers/network/utils/iptables/testing"
 	"k8s.io/kubernetes/pkg/features"
@@ -44,8 +46,6 @@ import (
 	proxyutiliptables "k8s.io/kubernetes/pkg/proxy/util/iptables"
 	proxyutiltest "k8s.io/kubernetes/pkg/proxy/util/testing"
 	"k8s.io/kubernetes/pkg/util/async"
-	utilipset "k8s.io/kubernetes/pkg/util/ipset"
-	ipsettest "k8s.io/kubernetes/pkg/util/ipset/testing"
 	utilipvs "k8s.io/kubernetes/pkg/util/ipvs"
 	ipvstest "k8s.io/kubernetes/pkg/util/ipvs/testing"
 	"k8s.io/utils/exec"
