--- pkg/util/ipset/OWNERS
@@ -1,13 +1,0 @@
-# See the OWNERS docs at https://go.k8s.io/owners
-
-reviewers:
-  - thockin
-  - brendandburns
-  - m1093782566
-  - islinwb
-approvers:
-  - thockin
-  - brendandburns
-  - m1093782566
-labels:
-- sig/network
